🐛 fix(curveframes): resolve `tau_unit` before touching the parameter
Review fixes, plus the array fastpath the previous commit documented but
left unreachable.

`FrenetSerretBuilder` called `.astype(float)` on the parameter *before*
resolving `tau_unit`, so a unitless parameter reported the accident
(`AttributeError: 'float' object has no attribute 'astype'`) instead of the
cause. Both sites now resolve first. `BishopBuilder` already did, which is
why only one of the two was caught by test.

The raw parameter then had nowhere to go: `unxt.experimental.jacfwd` strips
with `ustrip(unit, arg)`, which has no method for a bare array, so the array
fastpath died with a `NotFoundLookupError` several frames down -- declaring
`tau_unit` did not help, though the docs said it was what such a parameter
needed. A raw parameter is now wrapped with the declared unit at `_param`,
the single funnel every accessor takes its curve parameter from, so
`location`, `tangent`, `rotation_matrix`, `__call__` and a pinned raw
`station` all accept one, each agreeing with the `Quantity` call to the bit
and surviving `filter_jit`.

Note this wrap converges on the same `Quantity` computation immediately, so
it buys correctness, not speed: warm under `filter_jit` the array route
measures 0.079 ms/call against the `Quantity` route's 0.048. A fastpath that
is actually faster has to stay unitless end to end, which is a separate
change.

The `Typical usage::` block showed `curve[, tau_unit]`, which is not Python;
it now shows both call forms.

+++ packages/coordinaxs.curveframes/src/coordinaxs/curveframes/_src/frenetserret.py
@@ -173,8 +173,13 @@ class FrenetSerretBuilder(AbstractCurveFrameBuilder):
 
         # Unit-aware first and second derivatives via unxt. `g` is built
         # first: it is what the parameter unit is read off when undeclared.
-        g = b._param(p).astype(float)
+        # Resolve the unit *before* `.astype`: a unitless parameter has no
+        # `.astype`, so doing it the other way round reports the accident
+        # (`AttributeError`) instead of the cause (`_tau_unit_at`'s message,
+        # which names both ways to fix it).
+        g = b._param(p)
         tau_unit = b._tau_unit_at(g)
+        g = g.astype(float)
         dcurve = u.experimental.jacfwd(b.curve, units=(tau_unit,))
         d2curve = u.experimental.jacfwd(dcurve, units=(tau_unit,))
 
@@ -219,8 +224,9 @@ class FrenetSerretBuilder(AbstractCurveFrameBuilder):
 
         """
         b, p = self._resolve(tau)
-        g = b._param(p).astype(float)
+        g = b._param(p)
         dcurve = u.experimental.jacfwd(b.curve, units=(b._tau_unit_at(g),))
+        g = g.astype(float)
         return u.Q(_normalize(dcurve(g)).value, "")
 
     def normal(self, tau: Any, /) -> u.Q:
